pipelines/real_time/consumer/requirements.txt
influxdb-client==1.43.0
kafka-python==2.0.2